requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-identity
azure-mgmt-web
azure-keyvault-secrets
azure-storage-blob
azure-storage-queue
azure-cosmos

sendgrid
openpyxl
pandas
pymsteams
requests
json2html
pytz

# mysql-connector-python
# text2num
# pydantic
